#FEM-1136
Separated `AVPlayerEngine` into extension for better maintainability and readability.
